+++ crates/icydb-core/src/scalar_registry.rs
@@ -26,6 +26,9 @@ macro_rules! scalar_registry_family {
     (Bool) => {
         $crate::value::CoercionFamily::Bool
     };
+    (Unit) => {
+        $crate::value::CoercionFamily::Unit
+    };
 }
 
 #[doc(hidden)]